youtube_project/requirements.txt
requests
pymongo
pandas
python-dotenv
